Fix usage of Period instances
To get the seconds value of a period, it first has to be converted to a
Duration. Otherwise the wrong value (0) will be used when the Period is not
"PTnS" but "PTnH", for example.

Change updateInterval in CollectorRegistrationConfiguration from int to
long to avoid a type cast.

Fixes #57

+++ src/main/java/org/graylog/plugins/collector/collectors/rest/models/responses/CollectorRegistrationConfiguration.java
@@ -25,13 +25,13 @@ import com.google.auto.value.AutoValue;
 @JsonAutoDetect
 public abstract class CollectorRegistrationConfiguration {
     @JsonProperty
-    public abstract int updateInterval();
+    public abstract long updateInterval();
 
     @JsonProperty
     public abstract boolean sendStatus();
 
     @JsonCreator
-    public static CollectorRegistrationConfiguration create(@JsonProperty("update_interval") int updateInterval,
+    public static CollectorRegistrationConfiguration create(@JsonProperty("update_interval") long updateInterval,
                                                             @JsonProperty("send_status") boolean sendStatus) {
         return new AutoValue_CollectorRegistrationConfiguration(updateInterval, sendStatus);
     }
